requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.31.0
tenacity==8.3.0
pydantic==2.7.0
shapely==2.0.4
openpyxl==3.1.2
python-multipart==0.0.9
python-dotenv==1.0.1
instaloader==4.10.3

# KSA Merchant Scraper
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
pandas==2.2.2
tqdm==4.66.4
playwright==1.49.0
playwright-stealth==1.0.6
